split barebone in 3 files
